Implemented new Settings System into the config.
The Settings system allows the config to transfer data into the gui or
other systems.
This is done by creating a SettingsObject that can be in any format and
or Data Type as long it extends the interface.
The gui provides the implementations and the information provided for
the gui itself and tests for it.